requirements_api.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
httpx==0.27.0
redis==5.0.4
aioredis==2.0.1
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
shap==0.45.0
joblib==1.4.2
scikit-learn==1.4.2
xgboost==2.0.3
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0
prophet==1.1.5
